Begin translating to ReasonML
This leaves some issues, most notably in the fact that bs-fetch doesn't
happily support errors. This will need more work in the future.

Besides rewriting the components in JS this commit adds the required
BuckleScript packages (and bs-fetch for now) and converts the
styled-components set-up to a utility based set-up (inspired by
tailwind).

Webpack and scripts are reconfigured to use Reason as entry point
instead of the JS version.