Clean up JSON response traversing
This cleans up the accessing of the JSON response a bit with some help
from the ReasonML discord.

A good article on how this works can be found here:
https://medium.com/astrocoders/handling-optional-values-in-reasonml-with-belt-option-7e907819074b

Another good option to look into when doing more with Twitch is to build
a ReasonML library for Twitch that uses Decco to provide
encoders/decoders for Twitch API objects.

https://github.com/reasonml-labs/decco/