Implement Websocket server for scene control
This implements a websocket server in both production and development
that can control what is shown in the scene. This moves it from an
dumb client side application to something that can also accept webhooks
for things like notifications or chat messages.

This implements limited shared code to control what scene is displayed.
The scene is chosen on an interval to demonstrate that the ReScript code
that is shared between server and client works.

To ensure that the client and server data model remains in sync the
JavaScript code on the server may only write primitive data types such
as strings and integers and may not itself modify more complex types
(like the scene Variant).

We choose for a variant rather than a polymorphic variant because we
know what scenes are supported and we don't want this to be set to a
value that doesn't actually correspond to a scene (nor do we want to be
able to select a scene that is not mapped to a React component).