Convert to monorepo set-up
This converts the application to a monorepo set-up creating separate
packages for the existing server, viewer and shared code. This allows us
to add an admin package too for the stream control panel at `/admin/`.
This allows us to reuse module names between the client and server
without creating conflicts (e.g. having `Index.res` for each).

While creating a mono-repo set-up we convert to using `esbuild` for our
bundling since our needs are relatively simple. Additionally configuring
Webpack can be very tedious and as a bonus esbuild is a lot faster so we
compile all our packages synchronously in about 3-4 seconds. That should
lead to really quick deploy times.