Merge viewer and admin back into a single app
The most important change is that we now redirect all our requests to
our client application except for those that are specific to
authentication or are specifically in `/assets/`.

This allows us to do client side routing. This was previously not
possible but was attempted by splitting the client side applications.