Implement client side routing
This uses the ReactRouter to add some basic client side routing to
different pages. We take inspiration from Next.js to build our own Link
component that can be used to handle both internal and external links.

The StateMachine type is renamed to StreamState since we want to build
server side state that we do not expose to the client (such as the
secret control URLs).