Replace random scene switching with control panel
Implements separate websocket servers for the controller and the viewer.
Implements cookie based authentication for the viewer that stores the
result of the Twitch OAuth authentication. An authenticated controller
client can now change the scenes of the viewer. The viewer connection is
based on the URL.

Currently the controller connection will update global state rather than
client specific state. The viewer URL is also still hardcoded and not
client specific yet. These are things that need to be added in the
future.